requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
tldextract>=5.1.0
dnspython>=2.4.0
shodan>=1.30.0
censys>=2.2.0
openai>=1.0.0
markdown2>=2.4.0
pdfkit>=1.0.0
rich>=13.0.0
aiohttp>=3.8.0
pyyaml>=6.0
json5>=0.9.0
urllib3>=1.26.0
